Refactor type
