+++ packages/editor/src/models/ForeignComponentManager.ts
@@ -3,36 +3,7 @@ import { action, observable, reaction } from "mobx";
 import type React from "react";
 import type ReactDOM from "react-dom/client";
 import { projectState } from "../state/ProjectState";
-
-export type Type =
-  | {
-      type: "string";
-    }
-  | {
-      type: "boolean";
-    }
-  | {
-      type: "enum";
-      values: string[];
-    };
-
-export interface Prop {
-  name: string;
-  type: Type;
-}
-
-export interface ForeignComponent {
-  framework: "react"; // TODO: support other frameworks
-  path: string; // path relative to project root e.g. "src/Button.tsx"
-  name: string; // export name; e.g. "Button" ("default" for default export)
-  props: Prop[];
-  createRenderer: (element: HTMLElement) => ForeignComponentRenderer;
-}
-
-export interface ForeignComponentRenderer {
-  render(props: Record<string, unknown>): Promise<void>;
-  dispose(): void;
-}
+import { ForeignComponent } from "../types/ForeignComponent";
 
 export function foreignComponentKey(ref: { path: string; name: string }) {
   return `${ref.path}#${ref.name}`;
